changed sampling rate-floar

+++ GetFIFO.py
@@ -116,7 +116,7 @@ for loop in range (TargetSampleNumber):
     accel_zout_scaled = accel_zout
     x_rotation=get_x_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
     y_rotation=get_y_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
-    time_stamp=start_time + loop*((time.time()-start_time)/TargetSampleNumber)
+    time_stamp=start_time + float(loop*((time.time()-start_time)/TargetSampleNumber))
 
     msg = json.JSONEncoder().encode({"d":{"measured_timestamp":time_stamp, "gyro_xout_scaled":gyro_xout_scaled, "gyro_yout_scaled":gyro_yout_scaled, "gyro_zout_scaled":gyro_zout_scaled, "accel_xout_scaled":accel_xout_scaled, "accel_yout_scaled":accel_yout_scaled, "accel_zout_scaled":accel_zout_scaled, "x_rotation":x_rotation, "y_rotation":y_rotation}})
     print "pub", time_stamp
